Apply changes from review and discussion on TS * moved ScheduleID * removed unnecessary counter from EventScheduling * Used Plugin type for ExecutorServiceSchedulers to make testing & mocking easier * Smaller fixes to debug messages
