+++ src/main/java/org/betonquest/betonquest/api/schedule/Schedule.java
@@ -3,7 +3,7 @@ package org.betonquest.betonquest.api.schedule;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
 import org.betonquest.betonquest.exceptions.ObjectNotFoundException;
 import org.betonquest.betonquest.id.EventID;
-import org.betonquest.betonquest.id.ScheduleID;
+import org.betonquest.betonquest.modules.schedule.ScheduleID;
 import org.bukkit.configuration.ConfigurationSection;
 
 import java.util.ArrayList;
