+++ src/main/java/org/betonquest/betonquest/modules/schedule/ScheduleID.java
@@ -1,9 +1,10 @@
-package org.betonquest.betonquest.id;
+package org.betonquest.betonquest.modules.schedule;
 
 import org.betonquest.betonquest.Instruction;
 import org.betonquest.betonquest.api.config.QuestPackage;
 import org.betonquest.betonquest.api.schedule.Schedule;
 import org.betonquest.betonquest.exceptions.ObjectNotFoundException;
+import org.betonquest.betonquest.id.ID;
 
 /**
  * ID identifying a {@link Schedule}
