+++ src/main/java/org/betonquest/betonquest/modules/schedule/impl/realtime/RealtimeScheduler.java
@@ -7,7 +7,7 @@ import org.betonquest.betonquest.api.schedule.CronSchedule;
 import org.betonquest.betonquest.id.EventID;
 import org.betonquest.betonquest.modules.schedule.LastExecutionCache;
 import org.betonquest.betonquest.modules.schedule.impl.ExecutorServiceScheduler;
-import org.bukkit.Bukkit;
+import org.bukkit.plugin.Plugin;
 
 import java.time.Instant;
 import java.time.ZoneId;
@@ -39,11 +39,12 @@ public class RealtimeScheduler extends ExecutorServiceScheduler<RealtimeSchedule
     /**
      * Create a new realtime scheduler and pass BetonQuest instance to it
      *
-     * @param betonQuestInstance BetonQuest instance
+     * @param plugin             plugin used for bukkit scheduling, should be BetonQuest instance!
+     * @param lastExecutionCache cache where the last execution times of a schedule are stored
      */
-    public RealtimeScheduler(final BetonQuest betonQuestInstance) {
-        super(betonQuestInstance);
-        this.lastExecutionCache = betonQuestInstance.getLastExecutionCache();
+    public RealtimeScheduler(final Plugin plugin, final LastExecutionCache lastExecutionCache) {
+        super(plugin);
+        this.lastExecutionCache = lastExecutionCache;
     }
 
     @Override
@@ -67,7 +68,7 @@ public class RealtimeScheduler extends ExecutorServiceScheduler<RealtimeSchedule
                 .filter(CronSchedule::shouldRunOnReboot).toList();
         LOG.debug("Found " + rebootSchedules.size() + " reboot schedules. They will be run on next server tick.");
         if (!rebootSchedules.isEmpty()) {
-            Bukkit.getScheduler().runTaskLater(betonQuestInstance, () -> rebootSchedules.forEach(schedule -> {
+            plugin.getServer().getScheduler().runTaskLater(plugin, () -> rebootSchedules.forEach(schedule -> {
                 LOG.debug(schedule.getId().getPackage(), "Schedule " + schedule.getId() + " runs its events...");
                 for (final EventID eventID : schedule.getEvents()) {
                     BetonQuest.event(null, eventID);
@@ -84,7 +85,7 @@ public class RealtimeScheduler extends ExecutorServiceScheduler<RealtimeSchedule
         final List<RealtimeSchedule> missedSchedules = listMissedSchedules();
         LOG.debug("Found " + missedSchedules.size() + " missed schedule runs that will be caught up.");
         if (!missedSchedules.isEmpty()) {
-            Bukkit.getScheduler().runTaskLater(betonQuestInstance, () -> {
+            plugin.getServer().getScheduler().runTaskLater(plugin, () -> {
                 LOG.debug("Running missed schedules to catch up...");
                 for (final RealtimeSchedule missed : missedSchedules) {
                     lastExecutionCache.cacheExecutionTime(missed.getId(), Instant.now());
