+++ src/main/java/org/betonquest/betonquest/modules/schedule/impl/simple/SimpleSchedule.java
@@ -2,7 +2,7 @@ package org.betonquest.betonquest.modules.schedule.impl.simple;
 
 import org.betonquest.betonquest.api.schedule.Schedule;
 import org.betonquest.betonquest.exceptions.InstructionParseException;
-import org.betonquest.betonquest.id.ScheduleID;
+import org.betonquest.betonquest.modules.schedule.ScheduleID;
 import org.bukkit.configuration.ConfigurationSection;
 
 import java.time.Instant;
